fix(paths): move files to better paths
